Break down Auth functionality in several namespaces
