+++ lib/auth/index.ts
@@ -1,0 +1,1 @@
+export { Auth } from "./auth";
